Throw error message when polling with no interval
When using the `--poll` argument without an interval, this error would
come from chokidar:

UnhandledPromiseRejectionWarning: RangeError [ERR_OUT_OF_RANGE]: The value of "interval" is out of range. It must be an integer. Received NaN

It was not super helpful in figuring out what happened since I wasn't
directly using chokidar in my project. I was using ts-node-dev. This
updates the code to throw a helpful error message instead.